relocate hover styling to separate file
